Implemented password change and login forgot passwork link, create account link

+++ php/ajax.php
@@ -4,7 +4,7 @@ require_once ('header_functions.php');
 
 $functions = array('test', 'currentUser', 'currentUserId', 'addUser', 'getAllPages', 'getAllUsers', 'getMedia', 'getPage',
     'addMedia', 'addPage', 'deletePage', 'deleteUser', 'login', 'createAccount', 'createWebsite', 'getWebsiteData',
-    'getPagesData','getUsersData','getAccountMedia', 'deleteUser', 'deletePage', 'addUser','addPage');
+    'getPagesData','getUsersData','getAccountMedia', 'deleteUser', 'deletePage', 'addUser','addPage', 'updateAccountPassword');
 
 if(isset($_POST['function']) && in_array($_POST['function'], $functions)){
     $_POST['function']();
@@ -54,6 +54,20 @@ function login(){
     if($success === true){
         $response['accountId'] = $account->accountId;
         echo json_encode($response);
+    }else{
+        echo "false";
+    }
+    die;
+}
+
+function updateAccountPassword(){
+    $success = false;
+    if (!empty($_POST['email']) && !empty($_POST['password'])) {
+        $email = filter_var($_POST['email'], FILTER_VALIDATE_EMAIL);
+        $success = Account::updatePassword($email, $_POST['password']);
+    }
+    if($success === true){
+        echo "true";
     }else{
         echo "false";
     }
@@ -117,7 +131,7 @@ function getUsersData(){
 
 function addPage(){
     $success = false;
-    if (!empty($_POST['websiteId']) || !empty($_POST['pageName'])|| !empty($_POST['file']) ) {
+    if (!empty($_POST['websiteId']) && !empty($_POST['pageName']) && !empty($_POST['file']) ) {
         $schema = "website".$_POST['websiteId'];
         $data = Website::addPage($schema, $_POST['pageName'], $_POST['file']);
         if ($data !== false) {
@@ -134,8 +148,8 @@ function addPage(){
 
 function addUser(){
     $success = false;
-    if (!empty($_POST['websiteId']) || !empty($_POST['firstName'])|| !empty($_POST['lastName'])
-        || !empty($_POST['password'])|| !empty($_POST['email'])|| !empty($_POST['type']) ) {
+    if (!empty($_POST['websiteId']) && !empty($_POST['firstName'])&& !empty($_POST['lastName'])
+        && !empty($_POST['password'])&& !empty($_POST['email'])&& !empty($_POST['type']) ) {
         $schema = "website".$_POST['websiteId'];
         $data = Website::addUser($schema, $_POST['firstName'], $_POST['lastName'], $_POST['password'], $_POST['email'], $_POST['type']);
         if ($data !== false) {
@@ -152,7 +166,7 @@ function addUser(){
 
 function deletePage(){
     $success = false;
-    if (!empty($_POST['websiteId']) || !empty($_POST['pageId'])) {
+    if (!empty($_POST['websiteId']) && !empty($_POST['pageId'])) {
         $schema = "website".$_POST['websiteId'];
         $data = Website::deletePageById($schema, $_POST['pageId']);
         if ($data !== false) {
@@ -169,7 +183,7 @@ function deletePage(){
 
 function deleteUser(){
     $success = false;
-    if (!empty($_POST['websiteId']) || !empty($_POST['userId'])) {
+    if (!empty($_POST['websiteId']) && !empty($_POST['userId'])) {
         $schema = "website".$_POST['websiteId'];
         $data = Website::deleteUserById($_POST['userId'],$schema);
         if ($data !== false) {
